+++ examples/minimum/data.tf
@@ -1,7 +1,39 @@
+data "aws_partition" "current" {}
+
 data "aws_ecs_cluster" "ecs" {
   cluster_name = local.supporting_resources_name
 }
 
 data "aws_kms_alias" "supporting_kms" {
   name = "alias/${local.supporting_resources_name}"
+}
+
+data "aws_vpc" "supporting" {
+  filter {
+    name   = "tag:Name"
+    values = [local.supporting_resources_name]
+  }
+}
+
+data "aws_subnets" "private" {
+  filter {
+    name   = "tag:Name"
+    values = ["${local.supporting_resources_name}*.pri.*"]
+  }
+}
+
+data "aws_subnet" "private" {
+  for_each = toset(data.aws_subnets.private.ids)
+  id       = each.value
+}
+
+data "aws_iam_policy_document" "ecs_assume_role_policy" {
+  statement {
+    actions = ["sts:AssumeRole"]
+
+    principals {
+      type        = "Service"
+      identifiers = ["ecs-tasks.amazonaws.com"]
+    }
+  }
 }