+++ examples/minimum/main.tf
@@ -22,6 +22,15 @@ module "ecs_service" {
     }
   ]
 
+  service_security_group_egress = [
+    {
+      from_port   = 0
+      to_port     = 0
+      protocol    = "-1"
+      cidr_blocks = ["0.0.0.0/0"]
+    }
+  ]
+
   tags = {
     Name               = local.name
     Environment        = "examples"